Rename `Debug` namespace to avoid collisions
